backend/requirements.txt
flask
flask-sqlalchemy
flask-cors
opencv-python-headless
numpy
Pillow
python-dateutil
mysql-connector-python
torch
torchvision
easydict
insightface
onnxruntime
python-dotenv
